+++ src/hop_cli/aws/api/ssm.clj
@@ -3,8 +3,7 @@
 ;; file, You can obtain one at http://mozilla.org/MPL/2.0/
 
 (ns hop-cli.aws.api.ssm
-  (:require [clojure.set :as set]
-            [clojure.string :as str]
+  (:require [clojure.string :as str]
             [com.grzm.awyeah.client.api :as aws]
             [hop-cli.aws.api.client :as aws.client]))
 
@@ -16,22 +15,19 @@
   [{:keys [project-name environment]} name]
   (format "/%s/%s/env-variables/%s" project-name environment name))
 
-(def param-name-conversion
-  {:Value :value
-   :Name :name})
+(defn- api-params->env-vars
+  [api-params]
+  (reduce (fn [env-vars api-param]
+            (conj env-vars {(keyword (api-name->name (:Name api-param)))
+                            (:Value api-param)}))
+          {}
+          api-params))
 
-(defn- api-param->param
-  [api-param]
-  (-> api-param
-      (select-keys (keys param-name-conversion))
-      (set/rename-keys param-name-conversion)
-      (update :name api-name->name)))
-
-(defn put-parameter
-  [{:keys [project-name environment kms-key-alias] :as config} opts {:keys [name value]}]
+(defn- put-env-var-as-parameter
+  [{:keys [project-name environment kms-key-alias] :as config} opts [k v]]
   (let [ssm-client (aws.client/gen-client :ssm config)
-        request {:Name (name->api-name config name)
-                 :Value value
+        request {:Name (name->api-name config (name k))
+                 :Value v
                  :Type "SecureString"
                  :Tier "Standard"
                  :KeyId kms-key-alias}
@@ -52,13 +48,13 @@
        :error-details result}
       {:success? true})))
 
-(defn put-parameters*
+(defn- put-env-vars-as-parameters*
   [config opts parameters]
   (loop [pending-params parameters
          completed-results []]
     (if (empty? pending-params)
       completed-results
-      (let [result (put-parameter config opts (first pending-params))]
+      (let [result (put-env-var-as-parameter config opts (first pending-params))]
         (if (and
              (not (:success? result))
              (= "ThrottlingException" (get-in result [:error-details :__type])))
@@ -68,17 +64,17 @@
             (recur pending-params completed-results))
           (recur (rest pending-params) (conj completed-results result)))))))
 
-(defn put-parameters
+(defn put-env-vars-as-parameters
   [config opts parameters]
-  (let [results (put-parameters* config opts parameters)]
+  (let [results (put-env-vars-as-parameters* config opts parameters)]
     (if (every? :success? results)
       {:success? true}
       {:success? false
        :error-details (filter (comp not :success?) results)})))
 
-(defn get-parameters
+(defn get-env-vars-from-parameters
   [{:keys [project-name environment] :as config}]
-  (loop [params []
+  (loop [env-vars {}
          next-token nil]
     (let [ssm-client (aws.client/gen-client :ssm config)
           request {:Path (format "/%s/%s/env-variables" project-name environment)
@@ -91,29 +87,33 @@
         (if-not (= "ThrottlingException" (:__type result))
           {:success? false
            :error-details {:result result
-                           :succesfully-got-params (count params)}}
+                           :succesfully-got-params (count env-vars)}}
           (do
             (println "SSM Rate limit exceeded. Retrying in 3s...")
             (Thread/sleep 3000)
-            (recur params next-token)))
-        (let [all-params (concat params (map api-param->param (:Parameters result)))]
+            (recur env-vars next-token)))
+        (let [all-params (into env-vars (api-params->env-vars (:Parameters result)))]
           (if-let [next-token (:NextToken result)]
             (recur all-params next-token)
             {:success? true
              :params all-params}))))))
 
-(defn delete-parameters
-  [config parameters]
+(def ^:const ssm-max-delete-param
+  "Maximum number of SSM parameters that can be deleted in a single request."
+  10)
+
+(defn delete-env-vars-parameters
+  [config env-vars]
   (let [ssm-client (aws.client/gen-client :ssm config)]
-    (loop [params parameters
+    (loop [params (map name (keys env-vars))
            deleted-params []
            invalid-params []]
       (if (empty? params)
         {:success? true
          :deleted-params deleted-params
          :invalid-params invalid-params}
-        (let [params-to-delete (take 10 params)
-              param-names (map #(name->api-name config (:name %)) params-to-delete)
+        (let [params-to-delete (take ssm-max-delete-param params)
+              param-names (map #(name->api-name config %) params-to-delete)
               request {:Names param-names}
               args {:op :DeleteParameters
                     :request request}
@@ -123,6 +123,6 @@
              :error-details {:result result
                              :deleted-params deleted-params
                              :invalid-params invalid-params}}
-            (recur (nthnext params 10)
+            (recur (nthnext params ssm-max-delete-param)
                    (into deleted-params (:DeletedParameters result))
                    (into invalid-params (:InvalidParameters result)))))))))