+++ src/hop_cli/aws/env_vars.clj
@@ -4,14 +4,17 @@
 
 (ns hop-cli.aws.env-vars
   (:require [babashka.fs :as fs]
+            [clojure.data :refer [diff]]
+            [clojure.set :as set]
             [clojure.string :as str]
             [hop-cli.aws.api.eb :as api.eb]
             [hop-cli.aws.api.ssm :as api.ssm]
+            [hop-cli.bootstrap.util :as bootstrap.util]
             [hop-cli.util.thread-transactions :as tht]
             [malli.core :as m])
   (:import (java.util Date)))
 
-(def string-env-vars-schema
+(def strings-env-vars-schema
   [:sequential
    [:and
     [:string {:min 1}]
@@ -20,47 +23,37 @@
 (def ^:const last-ssm-script-update-env-var
   "LAST_SSM_SCRIPT_ENV_UPDATE")
 
-(defn- string-env-var->env-var
-  [s]
-  (zipmap [:name :value] (str/split s #"=" 2)))
+(def quoted-env-var-pattern
+  (re-pattern #"^'.*'$"))
 
-(defn- env-var->string-env-var
-  [{:keys [name value]}]
-  (let [env-file-quoted-val (-> value
-                                (str/replace "\\" "\\\\")
-                                (str/replace "'" "\\'"))]
-    (format "%s='%s'" name env-file-quoted-val)))
+(defn- string-env-vars->env-vars
+  [lines]
+  (reduce (fn [env-vars line]
+            (let [[k v] (str/split line #"=" 2)
+                  v (if (re-matches quoted-env-var-pattern v)
+                      (subs v 1 (dec (count v)))
+                      v)]
+              (assoc env-vars (keyword k) v)))
+          {}
+          lines))
 
-(defn- get-env-var-diff
+(defn- get-env-vars-diff
   [ssm-env-vars file-env-vars]
-  (let [all-env-var-names (distinct (map :name (concat ssm-env-vars file-env-vars)))]
-    (reduce
-     (fn [r name]
-       (let [in-ssm (some #(when (= name (:name %)) %) ssm-env-vars)
-             in-file (some #(when (= name (:name %)) %) file-env-vars)]
-         (cond
-           (and in-file (not in-ssm))
-           (update r :to-create conj in-file)
-
-           (and in-ssm (not in-file))
-           (update r :to-delete conj in-ssm)
-
-           (and in-file in-ssm (not= (:value in-ssm) (:value in-file)))
-           (update r :to-update conj in-file)
-
-           :else
-           r)))
-     {:to-update []
-      :to-create []
-      :to-delete []}
-     all-env-var-names)))
+  (let [keys-in-both (set/intersection (set (keys ssm-env-vars)) (set (keys file-env-vars)))
+        [different-in-file different-in-ssm _] (diff file-env-vars ssm-env-vars)
+        keys-to-create (set/difference (set (keys different-in-file)) keys-in-both)
+        keys-to-update (set/intersection (set (keys different-in-file)) keys-in-both)
+        keys-to-delete (set/difference (set (keys different-in-ssm)) keys-in-both)]
+    {:to-create (select-keys file-env-vars keys-to-create)
+     :to-update (select-keys file-env-vars keys-to-update)
+     :to-delete (select-keys ssm-env-vars keys-to-delete)}))
 
 (defn- sync-env-vars*
   [config {:keys [to-update to-create to-delete]}]
   (->
    [{:txn-fn
      (fn txn-1 [_]
-       (let [result (api.ssm/put-parameters config {:new? true} to-create)]
+       (let [result (api.ssm/put-env-vars-as-parameters config {:new? true} to-create)]
          (if (:success? result)
            {:success? true}
            {:success? false
@@ -68,13 +61,13 @@
             :error-details result})))
      :rollback-fn
      (fn rollback-1 [prv-result]
-       (let [result (api.ssm/delete-parameters config to-create)]
+       (let [result (api.ssm/delete-env-vars-parameters config to-create)]
          (when-not (:success? result)
            (prn "Create parameters rollback failed"))
          prv-result))}
     {:txn-fn
      (fn txn-2 [_]
-       (let [result (api.ssm/put-parameters config {} to-update)]
+       (let [result (api.ssm/put-env-vars-as-parameters config {} to-update)]
          (if (:success? result)
            {:success? true}
            {:success? false
@@ -86,7 +79,7 @@
        prv-result)}
     {:txn-fn
      (fn txn-2 [_]
-       (let [result (api.ssm/delete-parameters config to-delete)]
+       (let [result (api.ssm/delete-env-vars-parameters config to-delete)]
          (if (:success? result)
            {:success? true}
            {:success? false
@@ -114,35 +107,34 @@
 
 (defn sync-env-vars
   [{:keys [file] :as config}]
-  (if-let [string-env-vars (read-env-vars-file file)]
-    (if-not (m/validate string-env-vars-schema string-env-vars)
+  (if-let [strings-env-vars (read-env-vars-file file)]
+    (if-not (m/validate strings-env-vars-schema strings-env-vars)
       {:success? false
        :reason :file-contains-invalid-environment-variable-format
-       :error-details (m/explain string-env-vars-schema string-env-vars)}
-      (let [result (api.ssm/get-parameters config)]
+       :error-details (m/explain strings-env-vars-schema strings-env-vars)}
+      (let [result (api.ssm/get-env-vars-from-parameters config)]
         (if-not (:success? result)
           {:success? false
            :reason :could-not-get-ssm-env-vars
            :error-details result}
           (let [ssm-env-vars (:params result)
-                file-env-vars (map string-env-var->env-var string-env-vars)
-                env-var-diff (get-env-var-diff ssm-env-vars file-env-vars)
-                result (sync-env-vars* config env-var-diff)]
-            (assoc result :sync-details env-var-diff)))))
+                file-env-vars (string-env-vars->env-vars strings-env-vars)
+                env-vars-diff (get-env-vars-diff ssm-env-vars file-env-vars)
+                result (sync-env-vars* config env-vars-diff)]
+            (assoc result :sync-details env-vars-diff)))))
     {:success? false
      :error :file-not-found}))
 
 (defn download-env-vars
   [{:keys [file] :as config}]
-  (let [result (api.ssm/get-parameters config)]
+  (let [result (api.ssm/get-env-vars-from-parameters config)]
     (if-not (:success? result)
       {:success? false
        :reason :could-not-get-ssm-env-vars
        :error-details result}
       (let [result (->> (:params result)
-                        (map env-var->string-env-var)
-                        sort
-                        (fs/write-lines file))]
+                        (map bootstrap.util/env-var->string-env-var)
+                        (bootstrap.util/write-variables-to-file! file))]
         {:success? (boolean result)}))))
 
 (defn apply-env-var-changes