+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file. This change
 
 ## [UNRELEASED]
 - Various profiles: Backported a bunch of utility functions from various customer projects.
+- [core profile] Correctly handle environment variables with '$' (and other potentially problematic characters) in them [issue #24]
 
 ## [0.1.14] - 2024-11-20
 - Upgrade external dependencies of multiple libraries